removed the Utility folder/namespace as it made no sense that 2 things were in here but the rest were not
