+++ src/SevenDigital.Api.Wrapper.Integration.Tests/Utility/Http/HttpClientTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using NUnit.Framework;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
 
 namespace SevenDigital.Api.Wrapper.Integration.Tests.Utility.Http
 {
