+++ src/SevenDigital.Api.Wrapper.Unit.Tests/Exceptions/ExceptionFactoryTests.cs
@@ -3,8 +3,8 @@ using System.Net;
 using NUnit.Framework;
 using SevenDigital.Api.Schema;
 using SevenDigital.Api.Wrapper.Exceptions;
-using SevenDigital.Api.Wrapper.Utility.Http;
-using SevenDigital.Api.Wrapper.Utility.Serialization;
+using SevenDigital.Api.Wrapper.Http;
+using SevenDigital.Api.Wrapper.Serialization;
 
 namespace SevenDigital.Api.Wrapper.Unit.Tests.Exceptions
 {
