+++ src/SevenDigital.Api.Wrapper.Unit.Tests/FluentAPITests.cs
@@ -6,8 +6,8 @@ using NUnit.Framework;
 using SevenDigital.Api.Wrapper.EndpointResolution;
 using SevenDigital.Api.Schema;
 using System.Threading;
-using SevenDigital.Api.Wrapper.Unit.Tests.Utility.Http;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
+using SevenDigital.Api.Wrapper.Unit.Tests.Http;
 
 namespace SevenDigital.Api.Wrapper.Unit.Tests
 {
