+++ src/SevenDigital.Api.Wrapper.Unit.Tests/Serialization/ResponseParserTests.cs
@@ -2,10 +2,10 @@ using System;
 using System.Net;
 using NUnit.Framework;
 using SevenDigital.Api.Wrapper.Exceptions;
-using SevenDigital.Api.Wrapper.Utility.Http;
-using SevenDigital.Api.Wrapper.Utility.Serialization;
+using SevenDigital.Api.Wrapper.Http;
+using SevenDigital.Api.Wrapper.Serialization;
 
-namespace SevenDigital.Api.Wrapper.Unit.Tests.Utility.Serialization
+namespace SevenDigital.Api.Wrapper.Unit.Tests.Serialization
 {
 	[TestFixture]
 	public class ResponseParserTests
