+++ src/SevenDigital.Api.Wrapper.Unit.Tests/SevenDigital.Api.Wrapper.Unit.Tests.csproj
@@ -67,14 +67,14 @@
     <Compile Include="FluentAPITests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReleaseEndpoint\ReleaseEndpointTests.cs" />
-    <Compile Include="Utility\Http\FakeHttpClient.cs" />
-    <Compile Include="Utility\Http\NameValueExtensionsTests.cs" />
-    <Compile Include="Utility\Http\RequestCoordinatorTests.cs" />
-    <Compile Include="Utility\Http\DictionaryExtensionsTests.cs" />
-    <Compile Include="Utility\Serialization\ApiResponseDetectorTests.cs" />
-    <Compile Include="Utility\Serialization\ResponseParserTests.cs" />
-    <Compile Include="Utility\Serialization\StringDeserializerTests.cs" />
-    <Compile Include="Utility\Serialization\TestObject.cs" />
+    <Compile Include="Http\FakeHttpClient.cs" />
+    <Compile Include="Http\NameValueExtensionsTests.cs" />
+    <Compile Include="Http\RequestCoordinatorTests.cs" />
+    <Compile Include="Http\DictionaryExtensionsTests.cs" />
+    <Compile Include="Serialization\ApiResponseDetectorTests.cs" />
+    <Compile Include="Serialization\ResponseParserTests.cs" />
+    <Compile Include="Serialization\StringDeserializerTests.cs" />
+    <Compile Include="Serialization\TestObject.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SevenDigital.Api.Schema\SevenDigital.Api.Schema.csproj">
