+++ src/SevenDigital.Api.Wrapper/EndpointResolution/RequestCoordinator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SevenDigital.Api.Wrapper.EndpointResolution.OAuth;
 using SevenDigital.Api.Wrapper.EndpointResolution.RequestHandlers;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
 
 namespace SevenDigital.Api.Wrapper.EndpointResolution
 {
