+++ src/SevenDigital.Api.Wrapper/EndpointResolution/RequestHandlers/GetRequestHandler.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using SevenDigital.Api.Wrapper.EndpointResolution.OAuth;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
 
 namespace SevenDigital.Api.Wrapper.EndpointResolution.RequestHandlers
 {
