+++ src/SevenDigital.Api.Wrapper/Exceptions/ApiException.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Net;
 using System.Runtime.Serialization;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
 
 namespace SevenDigital.Api.Wrapper.Exceptions
 {
