+++ src/SevenDigital.Api.Wrapper/Exceptions/InvalidResourceException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.Serialization;
 using SevenDigital.Api.Schema;
-using SevenDigital.Api.Wrapper.Utility.Http;
+using SevenDigital.Api.Wrapper.Http;
 
 namespace SevenDigital.Api.Wrapper.Exceptions
 {
