+++ src/SevenDigital.Api.Wrapper/Extensions/HasArtistIdParameterExtensions.cs
