+++ src/SevenDigital.Api.Wrapper/Http/GzipHttpClient.cs
@@ -6,7 +6,7 @@ using System.Net;
 using System.Text;
 using SevenDigital.Api.Wrapper.EndpointResolution;
 
-namespace SevenDigital.Api.Wrapper.Utility.Http
+namespace SevenDigital.Api.Wrapper.Http
 {
 	public class GzipHttpClient : IHttpClient
 	{
