+++ src/SevenDigital.Api.Wrapper/Http/Response.cs
@@ -2,7 +2,7 @@ using System;
 using System.Net;
 using System.Collections.Generic;
 
-namespace SevenDigital.Api.Wrapper.Utility.Http
+namespace SevenDigital.Api.Wrapper.Http
 {
 	[Serializable]
 	public class Response
