+++ src/SevenDigital.Api.Wrapper/Serialization/ApiResponseDetector.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace SevenDigital.Api.Wrapper.Utility.Serialization
+namespace SevenDigital.Api.Wrapper.Serialization
 {
 	public class ApiResponseDetector : IApiResponseDetector
 	{
