+++ src/SevenDigital.Api.Wrapper/Serialization/IResponseParser.cs
@@ -1,0 +1,9 @@
+﻿using SevenDigital.Api.Wrapper.Http;
+
+namespace SevenDigital.Api.Wrapper.Serialization
+{
+	public interface IResponseParser<out T>
+	{
+		T Parse(Response response);
+	}
+}
